Split react-component test into 2 parts, one using a NODE_ENV development and 1 using NODE_ENV production
